Extract design system components and tokens into dedicated folder
