+++ frontend/src/components/FiltersPanel/SelectCharacterType.tsx
@@ -5,7 +5,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "../ui/select";
+} from "../../design-system/components/select";
 
 const characterTypeOptions: { value: CharacterType | "any"; label: string }[] =
   [
