+++ frontend/src/design-system/components/CheckIcon.tsx
